Further reduce code duplication with OpenSSLStreamCipher
